refactored duplicate code
refactored duplicate code
