+++ CSV2Paper/update_checker.py
@@ -6,30 +6,21 @@ from os import path
 from tempfile import gettempdir
 from tkinter import *
 from tkinter import ttk
-
+from base_window import BaseWindow
 from files import __location__
 from user_settings import UserSettings
 from windows_style_button import WindowsButton
 from windows_title_bar_button import WindowsTitleBarButton
 
-class Updater:
+class Updater(BaseWindow):
     def __init__(self, base: Tk, user_settings: UserSettings, on_start=False):
+        super().__init__(user_settings = user_settings)
+
         self.base = base
         self.user_settings = user_settings
         self.update_popup = Toplevel(base, takefocus=True)
         self.update_popup.grab_set()
         self.headers = {'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/50.0.2661.102 Safari/537.36'}
-        
-        self.window_bg = None
-        self.widget_bg = None
-        self.fg = None
-        self.insert_bg = None
-        self.disabled_bg = 'gray80'
-        self.select_bg = None
-        self.folder_icon_file = None
-        self.up_arrow_icon_file = None
-        self.down_arrow_icon_file = None
-        self.set_colors()
 
         self.update_popup.update_idletasks()
         x = base.winfo_rootx()
@@ -183,19 +174,3 @@ class Updater:
         starty = event.y_root
 
         self.title_bar.bind('<B1-Motion>', move_window)
-
-    def set_colors(self):
-        if self.user_settings.dark_mode_enabled:
-            self.window_bg = 'gray15'
-            self.widget_bg = 'gray35'
-            self.fg = 'white'
-            self.insert_bg = 'white'
-            self.disabled_bg = 'gray20'
-            self.select_bg = 'gray30'
-        else:
-            self.window_bg = 'SystemButtonFace'
-            self.widget_bg = 'SystemWindow'
-            self.fg = 'SystemWindowText'
-            self.insert_bg = 'SystemWindowText'
-            self.disabled_bg = 'gray80'
-            self.select_bg = 'SystemWindow'
